CLeaned up testing suit
